+++ src/test/java/com/flab/funding/rest/MemberDeliveryAddressRestAdapterTest.java
@@ -1,7 +1,7 @@
-package com.flab.funding.member;
+package com.flab.funding.rest;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.flab.funding.infrastructure.adapters.input.data.request.MemberPaymentMethodRegisterRequest;
+import com.flab.funding.infrastructure.adapters.input.data.request.MemberDeliveryAddressRegisterRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -26,7 +26,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @SpringBootTest
 @Transactional
 @AutoConfigureMockMvc
-public class MemberPaymentMethodRestAdapterTest {
+public class MemberDeliveryAddressRestAdapterTest {
 
     @Autowired
     private MockMvc mockMvc;
@@ -42,18 +42,23 @@ public class MemberPaymentMethodRestAdapterTest {
     }
 
     @Test
-    public void registerPaymentMethod() throws Exception {
+    void registerDeliveryAddress() throws Exception {
         // given
-        MemberPaymentMethodRegisterRequest request = MemberPaymentMethodRegisterRequest.builder()
+        MemberDeliveryAddressRegisterRequest request = MemberDeliveryAddressRegisterRequest.builder()
                 .userKey("1L")
                 .isDefault(true)
-                .paymentNumber("3565 43")
+                .zipCode("01234")
+                .address("서울특별시 강서구")
+                .addressDetail("OO 아파트 xxx동 xxxx호")
+                .recipientName("홍길동")
+                .recipientPhone("010-1111-2222")
                 .build();
 
-        //when
+        System.out.println("objectMapper.writeValueAsString(request) = " + objectMapper.writeValueAsString(request));
 
+        //when
         //then
-        this.mockMvc.perform(post("/paymentMethods")
+        this.mockMvc.perform(post("/deliveryAddresses")
                         .content(objectMapper.writeValueAsString(request))
                         .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
@@ -61,16 +66,25 @@ public class MemberPaymentMethodRestAdapterTest {
                 .andDo(document("{class-name}/{method-name}",
                         requestFields(
                                 fieldWithPath("userKey").description("회원번호(외부용)"),
-                                fieldWithPath("isDefault").description("대표 결제 수단"),
-                                fieldWithPath("paymentNumber").description("카드 번호/계좌 번호")
+                                fieldWithPath("isDefault").description("대표 배송지 주소"),
+                                fieldWithPath("zipCode").description("우편번호"),
+                                fieldWithPath("address").description("도로명 주소"),
+                                fieldWithPath("addressDetail").description("상세 주소"),
+                                fieldWithPath("recipientName").description("받는 사람 이름"),
+                                fieldWithPath("recipientPhone").description("받는 사람 연락처")
                         ),
                         responseFields(
-                                fieldWithPath("paymentMethodKey").description("결제 수단 ID(외부용)"),
+                                fieldWithPath("deliveryAddressKey").description("배송지 주소 ID(외부용)"),
                                 fieldWithPath("userKey").description("회원번호(외부용)"),
-                                fieldWithPath("isDefault").description("대표 결제 수단"),
-                                fieldWithPath("paymentNumber").description("카드 번호/계좌 번호"),
+                                fieldWithPath("isDefault").description("대표 배송지 주소"),
+                                fieldWithPath("zipCode").description("우편번호"),
+                                fieldWithPath("address").description("도로명 주소"),
+                                fieldWithPath("addressDetail").description("상세 주소"),
+                                fieldWithPath("recipientName").description("받는 사람 이름"),
+                                fieldWithPath("recipientPhone").description("받는 사람 연락처"),
                                 fieldWithPath("createdAt").description("등록일자"),
                                 fieldWithPath("updatedAt").description("수정일자")
                         )));
     }
+
 }
