+++ src/test/java/com/flab/funding/rest/MemberPaymentMethodRestAdapterTest.java
@@ -1,7 +1,7 @@
-package com.flab.funding.member;
+package com.flab.funding.rest;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.flab.funding.infrastructure.adapters.input.data.request.MemberDeliveryAddressRegisterRequest;
+import com.flab.funding.infrastructure.adapters.input.data.request.MemberPaymentMethodRegisterRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -26,7 +26,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @SpringBootTest
 @Transactional
 @AutoConfigureMockMvc
-public class MemberDeliveryAddressRestAdapterTest {
+public class MemberPaymentMethodRestAdapterTest {
 
     @Autowired
     private MockMvc mockMvc;
@@ -42,23 +42,18 @@ public class MemberDeliveryAddressRestAdapterTest {
     }
 
     @Test
-    void registerDeliveryAddress() throws Exception {
+    public void registerPaymentMethod() throws Exception {
         // given
-        MemberDeliveryAddressRegisterRequest request = MemberDeliveryAddressRegisterRequest.builder()
+        MemberPaymentMethodRegisterRequest request = MemberPaymentMethodRegisterRequest.builder()
                 .userKey("1L")
                 .isDefault(true)
-                .zipCode("01234")
-                .address("서울특별시 강서구")
-                .addressDetail("OO 아파트 xxx동 xxxx호")
-                .recipientName("홍길동")
-                .recipientPhone("010-1111-2222")
+                .paymentNumber("3565 43")
                 .build();
 
-        System.out.println("objectMapper.writeValueAsString(request) = " + objectMapper.writeValueAsString(request));
-
         //when
+
         //then
-        this.mockMvc.perform(post("/deliveryAddresses")
+        this.mockMvc.perform(post("/paymentMethods")
                         .content(objectMapper.writeValueAsString(request))
                         .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
@@ -66,25 +61,16 @@ public class MemberDeliveryAddressRestAdapterTest {
                 .andDo(document("{class-name}/{method-name}",
                         requestFields(
                                 fieldWithPath("userKey").description("회원번호(외부용)"),
-                                fieldWithPath("isDefault").description("대표 배송지 주소"),
-                                fieldWithPath("zipCode").description("우편번호"),
-                                fieldWithPath("address").description("도로명 주소"),
-                                fieldWithPath("addressDetail").description("상세 주소"),
-                                fieldWithPath("recipientName").description("받는 사람 이름"),
-                                fieldWithPath("recipientPhone").description("받는 사람 연락처")
+                                fieldWithPath("isDefault").description("대표 결제 수단"),
+                                fieldWithPath("paymentNumber").description("카드 번호/계좌 번호")
                         ),
                         responseFields(
-                                fieldWithPath("deliveryAddressKey").description("배송지 주소 ID(외부용)"),
+                                fieldWithPath("paymentMethodKey").description("결제 수단 ID(외부용)"),
                                 fieldWithPath("userKey").description("회원번호(외부용)"),
-                                fieldWithPath("isDefault").description("대표 배송지 주소"),
-                                fieldWithPath("zipCode").description("우편번호"),
-                                fieldWithPath("address").description("도로명 주소"),
-                                fieldWithPath("addressDetail").description("상세 주소"),
-                                fieldWithPath("recipientName").description("받는 사람 이름"),
-                                fieldWithPath("recipientPhone").description("받는 사람 연락처"),
+                                fieldWithPath("isDefault").description("대표 결제 수단"),
+                                fieldWithPath("paymentNumber").description("카드 번호/계좌 번호"),
                                 fieldWithPath("createdAt").description("등록일자"),
                                 fieldWithPath("updatedAt").description("수정일자")
                         )));
     }
-
 }
