+++ src/test/java/com/flab/funding/rest/MemberRestAdapterTest.java
@@ -1,4 +1,4 @@
-package com.flab.funding.member;
+package com.flab.funding.rest;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.flab.funding.domain.model.MemberGender;
