+++ src/test/java/com/flab/funding/service/MemberServiceTest.java
@@ -1,0 +1,47 @@
+package com.flab.funding.service;
+
+import com.flab.funding.domain.model.Member;
+import com.flab.funding.domain.model.MemberGender;
+import com.flab.funding.domain.model.MemberLinkType;
+import com.flab.funding.domain.service.MemberService;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.time.LocalDate;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+@ExtendWith(SpringExtension.class)
+@SpringBootTest
+@Transactional
+public class MemberServiceTest {
+
+    @Autowired
+    private MemberService memberService;
+
+    @Test
+    public void 회원가입() throws Exception {
+        //given
+        Member member = Member.builder()
+                .linkType(MemberLinkType.NONE)
+                .email("Test@gmail.com")
+                .userName("홍길순")
+                .nickName("테스터")
+                .phoneNumber("010-1111-2222")
+                .gender(MemberGender.FEMALE)
+                .birthday(LocalDate.of(1998,1,30))
+                .password("")
+                .build();
+
+        //when
+        Member savedMember = memberService.registMember(member);
+        Member findMember = memberService.getMemberByUserKey(savedMember.getUserKey());
+
+        //then
+        assertEquals(savedMember, findMember);
+    }
+}
